Reorganize storage
