+++ src/gateways/storage/private/match.ts
@@ -1,6 +1,7 @@
 import { pg } from './pg'
+import { TABLE as PODS_MATCHES } from './podsMatches'
 
-const table = 'matches'
+export const TABLE = 'matches'
 
 export interface MatchRecord {
   id: number
@@ -20,58 +21,54 @@ export interface MatchRecord {
   deadline?: Date
 }
 
+export type MatchRecordWithPodId = MatchRecord & { podId: number }
+
+const matchRecordWithPodIdColumns = [
+  `${TABLE}.id as id`,
+  `${TABLE}.createdAt as createdAt`,
+  `${TABLE}.updatedAt as updatedAt`,
+  `${TABLE}.playerAId as playerAId`,
+  `${TABLE}.playerBId as playerBId`,
+  `${TABLE}.winnerId as winnerId`,
+  `${TABLE}.firstPlayerId as firstPlayerId`,
+  `${TABLE}.victoryConditionId as victoryConditionId`,
+  `${TABLE}.deckAClanId as deckAClanId`,
+  `${TABLE}.deckARoleId as deckARoleId`,
+  `${TABLE}.deckASplashId as deckASplashId`,
+  `${TABLE}.deckBClanId as deckBClanId`,
+  `${TABLE}.deckBRoleId as deckBRoleId`,
+  `${TABLE}.deckBSplashId as deckBSplashId`,
+  `${TABLE}.deadline as deadline`,
+  `${PODS_MATCHES}.podId as podId`,
+]
+
 export async function insertMatch(
   match: Omit<MatchRecord, 'id' | 'createdAt' | 'updatedAt'>
 ): Promise<MatchRecord> {
-  return pg(table)
+  return pg(TABLE)
     .insert(match, '*')
     .then(([row]) => row)
 }
 
-export async function connectMatchToPod(matchId: number, podId: number): Promise<void> {
-  return pg('pods_matches').insert({ podId, matchId })
-}
-
-export type MatchRecordWithPodId = MatchRecord & { podId: number }
-
-const matchRecordWithPodIdColumns = [
-  'matches.id as id',
-  'matches.createdAt as createdAt',
-  'matches.updatedAt as updatedAt',
-  'matches.playerAId as playerAId',
-  'matches.playerBId as playerBId',
-  'matches.winnerId as winnerId',
-  'matches.firstPlayerId as firstPlayerId',
-  'matches.victoryConditionId as victoryConditionId',
-  'matches.deckAClanId as deckAClanId',
-  'matches.deckARoleId as deckARoleId',
-  'matches.deckASplashId as deckASplashId',
-  'matches.deckBClanId as deckBClanId',
-  'matches.deckBRoleId as deckBRoleId',
-  'matches.deckBSplashId as deckBSplashId',
-  'matches.deadline as deadline',
-  'pods_matches.podId as podId',
-]
-
 export async function fetchMatchesForMultiplePods(
   podIds: number[]
 ): Promise<MatchRecordWithPodId[]> {
-  return pg(table)
-    .leftJoin('pods_matches', 'matches.id', 'pods_matches.matchId')
-    .whereIn('pods_matches.podId', podIds)
+  return pg(TABLE)
+    .leftJoin(`${PODS_MATCHES}`, 'matches.id', `${PODS_MATCHES}.matchId`)
+    .whereIn(`${PODS_MATCHES}.podId`, podIds)
     .select(matchRecordWithPodIdColumns)
 }
 
 export async function fetchMatchesForMultipleParticipants(
   participantIds: number[]
 ): Promise<MatchRecord[]> {
-  return pg(table)
+  return pg(TABLE)
     .whereIn('playerAId', participantIds)
     .orWhereIn('playerBId', participantIds)
 }
 
 export async function fetchMatch(matchId: number): Promise<MatchRecord> {
-  return pg(table)
+  return pg(TABLE)
     .where('id', matchId)
     .first()
 }
@@ -89,7 +86,7 @@ export async function updateMatch(
     | 'deckBSplashId'
   >
 ): Promise<MatchRecord> {
-  const result = await pg(table)
+  const result = await pg(TABLE)
     .where('id', match.id)
     .update({ ...match, updatedAt: new Date() }, '*')
   return result[0]
