+++ src/gateways/storage/private/podsMatches.ts
@@ -1,0 +1,7 @@
+import { pg } from './pg'
+
+export const TABLE = 'pods_matches'
+
+export async function connectMatchToPod(matchId: number, podId: number): Promise<void> {
+  return pg(TABLE).insert({ podId, matchId })
+}
